requirements.txt
# Core dependencies
openai>=1.0.0
google-generativeai>=0.3.0
chromadb>=0.4.0
scikit-learn>=1.3.0
feedparser>=6.0.0
trafilatura>=1.6.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
tenacity>=8.2.0
numpy>=1.24.0

# Google Sheets integration
gspread>=5.0.0
google-auth>=2.0.0

# Streamlit web UI
streamlit>=1.28.0
plotly>=5.17.0

# Optional: Testing (개발 환경에서만 필요)
# pytest>=7.4.0


